fix(FEC-11971): when in PiP mode user is considered not engaged (#115)
**the issue:**
when in pip mode and switching tabs, user is not considered as "engaged" in realtime dashboard analytics.

**root cause:** 
realtime analytics dashboard is considering "user engagement" when playback is unmuted and tab is focused. there are another 2 scenarios but they include fullscreen mode, which is irrelevant in case of pip mode.
when switching to another tab, kava plugin is sending view event with tabMode=notFocused even when in pip mode. hence, the realtime graph is showing that the user is not engaged.

**solution:**
kava plugin will report tabMode=focused in case we are in pip mode.
Note- changed also viewabilityMode to "in view" in case we are in pip mode. up until now, the viewabilityMode was "not in view" even when entering pip.

Solves FEC-11971